Using redux-persistent to handle AsyncStorage data instead of getItem, setItem manually. This can help making code clean
