+++ package.json
@@ -26,6 +26,7 @@
     "react": "16.0.0-alpha.12",
     "react-native": "^0.47.0",
     "react-redux": "^5.0.6",
+    "redux-persist": "^4.10.0",
     "redux-thunk": "^2.2.0"
   }
 }
